final release

+++ cinema.cpp
@@ -1,6 +1,6 @@
 #include "cinema.h"
 
-cinema::cinema(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, double tot, string tC, string dC, double cC, unsigned int scC, bool hideC): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloC(tC), descrizioneC(dC), costoC(cC), schermicontC(scC), hdC(hideC) {}
+cinema::cinema(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, float tot, string tC, string dC, float cC, unsigned int scC, bool hideC): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloC(tC), descrizioneC(dC), costoC(cC), schermicontC(scC), hdC(hideC) {}
 
 cinema *cinema::clone() const
 {
@@ -12,7 +12,7 @@ string cinema::getTipoPacc() const
     return "Cinema";
 }
 
-double cinema::getTot() const
+float cinema::getTot() const
 {
     return totale;
 }
@@ -27,7 +27,7 @@ string cinema::getDescC() const
     return descrizioneC;
 }
 
-double cinema::getCostoC() const
+float cinema::getCostoC() const
 {
     return costoC;
 }
