+++ kids.cpp
@@ -1,7 +1,7 @@
 #include "kids.h"
 
 
-kids::kids(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, double tot, string tK, string dK, double cK, unsigned int scK, bool hideK): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloK(tK), descrizioneK(dK), costoK(cK), schermicontK(scK), hdK(hideK) {}
+kids::kids(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, float tot, string tK, string dK, float cK, unsigned int scK, bool hideK): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloK(tK), descrizioneK(dK), costoK(cK), schermicontK(scK), hdK(hideK) {}
 
 kids *kids::clone() const
 {
@@ -13,7 +13,7 @@ string kids::getTipoPacc() const
     return "Kids";
 }
 
-double kids::getTot() const
+float kids::getTot() const
 {
     return totale;
 }
@@ -28,7 +28,7 @@ string kids::getDescK() const
     return descrizioneK;
 }
 
-double kids::getCostoK() const
+float kids::getCostoK() const
 {
     return costoK;
 }
