+++ sport.cpp
@@ -1,6 +1,6 @@
 #include "sport.h"
 
-sport::sport(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, double tot, string tS, string dS, double cS, unsigned int scS, bool shdS): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloS(tS), descrizioneS(dS), costoS(cS), schermicontS(scS), superhdS(shdS) {}
+sport::sport(string n, string c, int aN, int mN, int gN, string cf, string em, int aA, int mA, int gA, unsigned int me, bool cosc, string coup, float tot, string tS, string dS, float cS, unsigned int scS, bool shdS): nowqt(n, c, aN, mN, gN, cf, em, aA, mA, gA, me, cosc, coup, tot), titoloS(tS), descrizioneS(dS), costoS(cS), schermicontS(scS), superhdS(shdS) {}
 
 sport *sport::clone() const
 {
@@ -12,7 +12,7 @@ string sport::getTipoPacc() const
     return "Sport";
 }
 
-double sport::getTot() const
+float sport::getTot() const
 {
     return totale;
 }
@@ -27,7 +27,7 @@ string sport::getDescS() const
     return descrizioneS;
 }
 
-double sport::getCostoS() const
+float sport::getCostoS() const
 {
     return costoS;
 }
